RobustFederation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.19.0=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
blis=0.7.9=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.3.1=pypi_0
catalogue=2.0.8=pypi_0
certifi=2022.9.24=py38h06a4308_0
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
confection=0.0.4=pypi_0
contourpy=1.0.7=pypi_0
cryptography=38.0.1=py38h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.7.99=0
cuda-gdb=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
decorator=5.1.1=pypi_0
einops=0.6.0=pypi_0
ema-pytorch=0.2.3=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.11.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.39.3=pypi_0
freetype=2.12.1=h4a9f257_0
ftfy=6.1.1=pypi_0
gdown=4.7.1=pypi_0
gds-tools=1.4.0.31=0
geom-median=0.1.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.20.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.2=pypi_0
h11=0.14.0=pypi_0
huggingface-hub=0.13.4=pypi_0
idna=3.4=py38h06a4308_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
kmeans-pytorch=0.3=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
librosa=0.10.1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
llvmlite=0.41.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
msgpack=1.0.7=pypi_0
murmurhash=1.0.9=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
nsight-compute=2022.3.0.22=0
numba=0.58.1=pypi_0
numpy=1.23.4=py38h14f4228_0
numpy-base=1.23.4=py38h31eccc5_0
nvidia-ml-py=12.535.161=pypi_0
nvitop=1.3.2=pypi_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1s=h7f8727e_0
outcome=1.3.0.post0=pypi_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pathy=0.10.1=pypi_0
patsy=0.5.4=pypi_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
platformdirs=4.0.0=pypi_0
pooch=1.8.0=pypi_0
preshed=3.0.8=pypi_0
prettytable=3.10.0=pypi_0
progress=1.6=pypi_0
protobuf=4.23.3=pypi_0
psutil=5.9.5=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.9=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pypi_0
pytorch=1.13.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-fid=0.3.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
readline=7.0=h7b6447c_5
regex=2023.6.3=pypi_0
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
selenium=4.17.2=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
smart-open=6.3.0=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.4.1=pypi_0
soxr=0.3.7=pypi_0
spacy=3.5.3=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.4=pypi_0
sqlite=3.33.0=h62c20be_0
srsly=2.4.6=pypi_0
statsmodels=0.14.0=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
termcolor=2.4.0=pypi_0
thinc=8.1.10=pypi_0
threadpoolctl=3.1.0=pypi_0
timm=0.6.13=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.13.0=py38_cu117
torchmetrics=0.11.4=pypi_0
torchvision=0.14.0=py38_cu117
tqdm=4.65.0=pypi_0
trio=0.24.0=pypi_0
trio-websocket=0.11.1=pypi_0
typer=0.7.0=pypi_0
typing-extensions=4.9.0=pypi_0
urllib3=1.26.12=py38h06a4308_0
wasabi=1.1.2=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wsproto=1.2.0=pypi_0
xgboost=2.0.2=pypi_0
xlwt=1.3.0=pypi_0
xz=5.2.8=h5eee18b_0
yacs=0.1.8=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
